requirements.txt
annotated-doc              0.0.4
annotated-types            0.7.0
anyio                      4.12.1
asn1crypto                 1.5.1
bcrypt                     3.2.2
boto3                      1.42.43
botocore                   1.42.43
certifi                    2026.1.4
cffi                       2.0.0
charset-normalizer         3.4.4
click                      8.3.1
colorama                   0.4.6
cryptography               46.0.4
dnspython                  2.8.0
email-validator            2.3.0
fastapi                    0.128.2
filelock                   3.20.3
greenlet                   3.3.1
h11                        0.16.0
idna                       3.11
Jinja2                     3.1.6
jmespath                   1.1.0
MarkupSafe                 3.0.3
packaging                  26.0
passlib                    1.7.4
pip                        25.2
platformdirs               4.5.1
pycparser                  3.0
pydantic                   2.12.5
pydantic_core              2.41.5
PyJWT                      2.11.0
PyMySQL                    1.1.2
pyOpenSSL                  25.3.0
python-dateutil            2.9.0.post0
python-multipart           0.0.22
pytz                       2025.2
requests                   2.32.5
s3transfer                 0.16.0
six                        1.17.0
snowflake-connector-python 4.2.0
sortedcontainers           2.4.0
SQLAlchemy                 2.0.46
starlette                  0.50.0
tomlkit                    0.14.0
typing_extensions          4.15.0
typing-inspection          0.4.2
urllib3                    2.6.3
uvicorn                    0.40.0